Time: 7 ms (12.21%) | Memory: 8.2 MB (67.12%) - LeetSync

+++ 8-string-to-integer-atoi/string-to-integer-atoi.cpp
@@ -5,7 +5,7 @@ public:
             return 0;
 
         int sign = 1;
-        long ans = 0; 
+        double ans = 0; 
         int i = 0;
 
         // Remove leading whitespace
